// frontend/src/api/models/CourseCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CourseCreate
 */
export interface CourseCreate {
    /**
     * 
     * @type {string}
     * @memberof CourseCreate
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof CourseCreate
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CourseCreate
     */
    categoryId: number;
    /**
     * 
     * @type {number}
     * @memberof CourseCreate
     */
    teacherId: number;
    /**
     * 
     * @type {number}
     * @memberof CourseCreate
     */
    deadlineInDays: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof CourseCreate
     */
    isActive?: boolean;
}

/**
 * Check if a given object implements the CourseCreate interface.
 */
export function instanceOfCourseCreate(value: object): value is CourseCreate {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('categoryId' in value) || value['categoryId'] === undefined) return false;
    if (!('teacherId' in value) || value['teacherId'] === undefined) return false;
    if (!('deadlineInDays' in value) || value['deadlineInDays'] === undefined) return false;
    return true;
}

export function CourseCreateFromJSON(json: any): CourseCreate {
    return CourseCreateFromJSONTyped(json, false);
}

export function CourseCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): CourseCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'categoryId': json['category_id'],
        'teacherId': json['teacher_id'],
        'deadlineInDays': json['deadline_in_days'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
    };
}

export function CourseCreateToJSON(json: any): CourseCreate {
    return CourseCreateToJSONTyped(json, false);
}

export function CourseCreateToJSONTyped(value?: CourseCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'description': value['description'],
        'category_id': value['categoryId'],
        'teacher_id': value['teacherId'],
        'deadline_in_days': value['deadlineInDays'],
        'is_active': value['isActive'],
    };
}